move watchpoint definitions around
